[Grunt] Added express, open, and watch tasks to improve dev/build
The express task enables a localhost server which is automatically run
when `Grunt` is run and sets the docroot to 'build'.

The Open task will automatically open up a browser tab at http://localhost:9000/build/example.

The Watch task will watch for any changes to the example source files and the ThemeEditor files
and then automatically lint, ugligy and copy the files to the build directory, and refresh them using livereload.