[theme-modifier] Added Two Way Data binding. Fix #4, Close #5.
This now allows inputs to declare a `data-cluckles-{{type}=“property”`
and the data binding will automatically be applied and managed by the
matching {{type}}.js file. Along as that js file has a `modifier` which
matches the `property`.

The DOM Element which has the binding will automatically update the
modifier and apply the change, aswell as be notified when the modifier
is changes (such as when using the API).