[ThemeEditor] Added TWDB to the JS editor files.
Added Two Way Data Binding to the JS files which make up chuckles.
This now means that the data binding that is attached to the editor
HTML templates is now accompanied by changes to the JS so it is now
completely functional.

All that it needs now is to trigger the binding and update the view
when theme files are loaded.